requirements.txt
openai>=1.0.0
faster-whisper>=1.0.0
opencv-python-headless>=4.8.0
moviepy==1.0.3
yt-dlp>=2024.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
assemblyai>=0.20.0
boto3>=1.34.0
